move package from server.decoding to server.encoding
