+++ core/src/main/java/com/linecorp/armeria/server/encoding/DecodingService.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.server.decoding;
+package com.linecorp.armeria.server.encoding;
 
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static java.util.Objects.requireNonNull;
