+++ core/src/main/java/com/linecorp/armeria/server/encoding/HttpDecodedRequest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.server.decoding;
+package com.linecorp.armeria.server.encoding;
 
 import org.reactivestreams.Subscriber;
